iommu/fsl: Use pci_clear_master to disable device bus master capability.
Initially we were disabling the bus master capability by writing to the
PCI configuration space. We were not clearing the "is_busmaster" flag.
As a result when the device was assgined back to host, the bus master
capability of the pci bridge (to which the device was connected) were
not getting enabled. Once the device was assigned back to host, device
was facing DMA issues.

Now, we are using the pci_clear_master for disabling device bus master
capability. This routine clears the device bus master capability by
accessing the device configuration space and also clears the
"is_busmaster" flag.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>
Change-Id: Icda6f5551d8da43bec8f2b657846b7e3f9474290
Reviewed-on: http://git.am.freescale.net:8181/12224
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>