staging: fsl-mc: Do programing of MSIs outside of devm_request_threaded_irq()
Normally, devm_request_threaded_irq() programs the MSI physically in
the device (by invoking a device-specific callback). However, for MC
IRQs, we have to program the MSI outside of this callback, because this
callback is invoked with interrupts disabled, and we don't have a reliable
way of sending commands to the MC from atomic context. This will be even
more so, once we add mutex locking to mc_send_command(). When invoked
from atomic context, mc_send_command() will to a mutex_trylock() and
if it is not able to grab the mutex, it will return with error without
sending the command to the MC.

Signed-off-by: J. German Rivera <German.Rivera@freescale.com>
Change-Id: I091b889f6f5056157c5e232dd633073a5dac49ca
Reviewed-on: http://git.am.freescale.net:8181/34640
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>