staging: fsl-mc: Add locking to serialize mc_send_command() calls
Add a locking mechanism to serialize mc_send_command() calls that use
the same fsl_mc_io object (same MC portal). All mc_send_command() calls
with the same fsl_m_io object have to be either from non-atomic context
or from atomic context, but not both. When the fsl_mc_io object is
created the owner needs to know in which type of context the fsl_mc_io
object is going to be used. A flag passed-in to fsl_create_mc_io()
will indicate whether the fsl_mc_io object will be used in atomic or
non-atomic context. If the fsl_mc_io object is going to be used in
non-atomic context, mc_send_command() calls with it will be
serialized using a mutex. Otherwise, if the fsl_mc_io object is
going to be used in atomic context, mc_semd_command() calls with it
will be serialized using a spinlock.

Signed-off-by: J. German Rivera <German.Rivera@freescale.com>
Change-Id: Icb770cd36e204ee6a17ad0f81e1d31cc9fe96816
Reviewed-on: http://git.am.freescale.net:8181/34876
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>