fsl-mc-dpio: Make QBMan token tracking internal
Previousy the QBMan portal code required the caller to properly set and
check for a token value used by the driver to detect when the QMan
hardware had completed a dequeue.  This patch simplifes the driver
interface by internally dealing with token values.  The driver will now
set the token value to 0 once it has dequeued a frame while a token
value of 1 indicates the HW has completed the dequeue but SW has not
consumed the frame yet.

Signed-off-by: Roy Pledge <Roy.Pledge@freescale.com>
Change-Id: If94d9728b0faa0fd79b47108f5cb05a425b89c18
Reviewed-on: http://git.am.freescale.net:8181/35433
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Haiying Wang <Haiying.Wang@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>