dpaa2-eth: Fix "NOHZ: local_softirq_pending" warning
Explicitly run softirqs after we enable NAPI. This in particular gets us
rid of the "NOHZ: local_softirq_pending" warnings, but it also solves a
couple of other problems, among which fluctuating performance and high
ping latencies.

Notes:
 - This will prevent us from timely processing notifications and
other "non-frame events" coming into the software portal. So far,
though, we only expect Dequeue Available Notifications, so this patch
is good enough for now.
 - A degradation in console responsiveness is expected, especially in
cases where the bottom-half runs on the same CPU as the console.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: Ia6f11da433024e80ee59e821c9eabfa5068df5e5
Reviewed-on: http://git.am.freescale.net:8181/35830
Reviewed-by: Alexandru Marginean <Alexandru.Marginean@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>
Tested-by: Stuart Yoder <stuart.yoder@freescale.com>