dpaa2-eth: limited support for flow steering
Steering is supported on a sub-set of fields, including DMAC, IP SRC
and DST, L4 ports.
Steering and hashing configurations depend on each other, that makes
the whole thing tricky to configure.  Currently FS can be configured
using only the fields selected for hashing and all the hashing fields
must be included in the match key - masking doesn't work yet.

Signed-off-by: Alex Marginean <alexandru.marginean@freescale.com>
Change-Id: I9fa3199f7818a9a5f9d69d3483ffd839056cc468
Reviewed-on: http://git.am.freescale.net:8181/38759
Reviewed-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Ruxandra Ioana Radulescu <ruxandra.radulescu@freescale.com>
Reviewed-by: Stuart Yoder <stuart.yoder@freescale.com>