dpaa2-eth: Allocate channels based on queue count
Limit the number of channels allocated per DPNI to the maximum
between the number of Rx queues per traffic class (distribution size)
and Tx confirmation queues (number of tx flows).
If this happens to be larger than the number of available cores, only
allocate one channel for each core and distribute the frame queues on
the cores/channels in a round robin fashion.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>