Avoid usage of preemptible() macro
The macro preemptible() was used in mc-sys.c to decide if
mc_send_command may sleep while waiting for MC command to complete.
Usage of this macro is not safe and thus another approach is taken.
Following the change, the decision to sleep or not is based on
the value of atomic flag in mc_io so it is caller's responssibility
to use the right portal.

Signed-off-by: Itai Katz <itai.katz@freescale.com>