dpaa2-eth: Iterate only through initialized channels
The number of DPIO objects available to a DPNI may be fewer than the
number of online cores. A typical example would be a DPNI with a
distribution size smaller than 8. Since we only initialize as many
channels (DPCONs) as there are DPIOs, iterating through all online cpus
would produce a nasty oops when retrieving ethtool stats.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>