KVM: PPC: e6500: TLB emulation for IND entries
Handle indirect entries (IND) in TLB emulation code.
Translation size of IND entries differs from the size
of referred Page Tables (Linux guests now use IND
of 2MB for 4KB PTs) and this requires careful tweaking
of the existing logic.

TLB search emulation requires additional search in
HW TLB0 (since entries might be directly added by HTW)
and found entries should be presented to the guest with
RPN changed from PFN to GFN. There might be more GFNs
pointing to the same PFN so the only way to get the
corresponding GFN is to search it in guest's PTE.
If the IND entry for the corresponding PT is not available
just invalidate guest's ea and report a tlbsx miss.
This patch only implements the invalidation and lets a
TODO note for searching HW TLB0.

Signed-off-by: Mihai Caraman <mihai.caraman@freescale.com>
[Laurentiu.Tudor@freescale.com: addressed review feedback,
     refactoring, cleanup & other fixes]
Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>