dts/ls2085a-rdb: Fix DPMAC connectivity to AQR PHYs on emdio1
Put the DTS in sync with the current DPMAC driver implementation, by
explicitly referring external PHYs from within the respective DPMAC nodes.

This configuration refers to SerDes 0x2a_0x41, which at the moment is the
only supported on RDB; this being the reason why the dts filename itself
bears no reference to the SerDes number.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>