powerpc/e6500: set the SPRN_PIR register only after e6500 core resets
If a core has reset and its threads have been initialized, the SPRN_PIR
registers have the correct value, and there is no need to initialize it
again. This patch checks the SPRN_BUCSR register to see if the core has
been reset. Such things will happen, for example, in the CPU hotplug case.

This patch derived from the following upstream patch:
commit f34b3e19fd06 ("powerpc/e6500: kexec: Handle hardware threads")

Signed-off-by: Chenhui Zhao <chenhui.zhao@nxp.com>