fsl-dpaa2: mac: Fix interrupt handling
The DPMAC has two interrupt events muxed on a single interrupt line.
Both the PHY and the DPNI can initiate a link event.

When the link event is initiated by the PHY (possibly as the effect of an
earlier link change request initiated by a DPNI), we must make sure
dpmac_set_link_state() is explicitly called in order for the event to be
propagated (back) to the DPNI.

Finally, DPMAC interrupt mask has to be explicitly specified before calling
dpmac_set_irq_enabled().

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>