dpaa2-eth: Force atomic context for lazy bpool seeding
We use the same ldpaa_bp_add_7() function for initial buffer pool
seeding (from .ndo_open) and for hotpath pool replenishing. The function
is using napi_alloc_frag() as an optimization for the Rx datapath, but
that turns out to require atomic execution because of a this_cpu_ptr()
call down its stack.
This patch temporarily disables preemption around the initial seeding of
the Rx buffer pool.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>