arm: pm: add deep sleep support for LS1
LS1 supports deep sleep feature that can switch off most parts
of the SoC when it is in deep sleep state.

The DDR controller will also be powered off in deep sleep.
Therefore, copy the last stage code to enter deep sleep to SRAM
and run it with disabling MMU and caches.

Signed-off-by: Chenhui Zhao <chenhui.zhao@nxp.com>
Signed-off-by: Wang Dongsheng <dongsheng.wang@nxp.com>