clocksource: arm_arch_timer: Timer reread erratum
Work around Freescale erratum A-008585 ("ARM Generic Timer may contain
an erroneous value"), which says that the TVAL and CNTnCT registers may
"contain an erroneous value for a small number of core clock cycles
every time the timer value changes".  The workaround is to read the
register twice, with back-to-back instructions, and repeat if the
values do not match.

This erratum is present on LS2080A (64-bit), LS1043A (64-bit), LS1021A
(32-bit) and derivatives.  This patch only addresses 64-bit.

This patch is loosely based on work by Priyanka Jain and Bhupesh
Sharma.

Signed-off-by: Scott Wood <scott.wood@nxp.com>
(Stuart: split out device tree change)
Signed-off-by: Stuart Yoder <stuart.yoder@nxp.com>