dpa_offload/dpa_stats: get_counters event handling redesign
Changes were made to the dpa_stats get_counters event handling
mechanism in order to support also the LE (Little Endian) host
architectures. The old implementation relied on PPC specific
instructions to determine wether US was compiled for 32/64 bit in
the case of a 64 bit kernel. A redesign was done in order to
circumvent this issue, thus removing the need to pass callbacks
between US and KS any more.

Signed-off-by: Rosioru Dragos Ionut-B41654 <dragos.rosioru@freescale.com>
Reviewed-by: Marian Chereji <marian.chereji@nxp.com>
Verified-by: Marian Chereji <marian.chereji@nxp.com>