crypto: caam - lower QI congestion prints verbosity level
Whenever congestion occurs on the SEC response FQs, a message
is printed on the console. The same happens when the congestion
no longer happens.
These prints, albeit rate limited, are detrimental to the overall
system performance.
Thus, this patch makes these prints to be shown only on the
debug error level, as well as add a new debugfs counter for
counting how many times the congestion state has occurred.
The debugfs file where this is held can be found in the
following path:
/sys/kernel/debug/<CAAM device name>/ctl/qi_congested

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>