arm64/config: enable HMAC-SHA(224|256|384|512) algorithms in kernel.
If an algorithm software implementation is not enabled in kernel,
then when requesting a transformation using that specific algorithm,
even if there's a hardware offload for that algorithm, the CryptoAPI
will return that the particular algorithm is not supported.
As an example, trying to create an IPSec tunnel using AES-CBC & HMAC-SHA384
without having CONFIG_CRYPTO_SHA512=y in the defconfig, will lead to a
message similar to the following:

<snip>
add 200.200.200.10 200.200.200.20 esp 0x10513
    -E aes-cbc        0x7aeaca3f87d060a12f4a4487d5a5c3355920fae69a96c831
    -A hmac-sha2-384  0x4de03bebf6beb4fdef5a67d349a09580466cc4e54503333b2a5fd34538c91198ffffffffffffffffffffffffffffffff;
<snip>

line xx: Not supported at [0x4de03bebf6beb4fdef5a67d349a09580466cc4e54503333b2a5fd34538c91198ffffffffffffffffffffffffffffffff]
parse failed, line xx.

This patch enables in the fragment file the algorithms mentioned above,
which in turn ensures that the accelerated counterparts can also be used.

It is worth mentioning that in the defconfig files for PPC platforms,
these algorithms are already activated.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>