fsl mc-bus: Move mc-io destroy to bus remove
fsl_mc_device_add() is called when a new mc device is added and
fsl_mc_device_remove() is called when a mc device is removed.
Now mc portals (fsl_create_mc_io) are not created during device
addition i.e fsl_mc_device_add(). But mc portal are being destroyed
when an mc device are removed i.e fsl_mc_device_remove(), which is
not correct way of reverting what is done during device addition.

Now this leads to the problem with VFIO driver where we do not want
the portal to be destroyed when an mc-device is removed. VFIO
creates one mc-portal for all dprc-containers bound to it and it
destroys the mc portal when driver itself is removed.

This change moves the mc-portal destroy part to bus-remove for
mc-bus and dprc-driver remove for child DPRCs. This also makes
fsl_mc_device_remove() exactly reverting what is done in
fsl_mc_device_add() with respect to mc portal creation.

Signed-off-by: Bharat Bhushan <Bharat.Bhushan@nxp.com>