dpa_offload: Fix buffer pool conflict for oNIC
The oNIC interface needs to have an exclusive buffer pool on its
offline port and, during the device tree files resynchronization,
the buffer pool from offline port #2 was changed to bpool #16. This
buffer pool is also used by different Ethernet interfaces that are
seeding it and, when used on oNIC, the driver would crash while
seeding the buffer pool a second time.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>