dpaa_ethernet: add two-tiered RX PCD FQs
In order to prioritize one type of RX traffic over
some other RX traffic, one can use a different range
of RX FQs for distributing the higher priority traffic
to the CPU cores.
This is very useful f.i. in back2back scenarios
used in IPSec applications.
For this to happen, it's needed to use a different range
of FQs for the high priority traffic such as existing
use-cases are not affected. This in turn requires increasing
the number of FQ IDs in the system. Furthermore, for
taking advantage of this mechanism, the corresponding XML
files also require updates.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>