fsl_dce/tests: Fix sample test data declarations
The common data used for DCE primitives tests was included in the
driver. This was not the correct location for test specific data so
it was moved out to be built with the test modules. The test modules
each include the sample data and if compiled as loadable modules no
issues appeared. However when the tests were built into the kernel the
sample data declarations conflicted. The sample data declarations are
now correctly declared as static to prevent the declaration conflicts

Signed-off-by: Ahmed Mansour <Ahmed.Mansour@nxp.com>