dpa_offload/dpa_classifier: Improve support for "return-to-keygen" table entries
The dpa_classifier tables support entries with a special type of
action called "return-to-keygen". This type of entries allow the
user to send selected network packets back to KeyGen for further
processing, such as extracting a new type of classification key.
The packets are directed to a special type of FMan distribution
called "direct scheme".

Up until now the dpa_classifier API only allowed the user to add
ONE single type of entries - either all the entries in the table
were of type "return-to-keygen", or all the entries in the table
were simple action entries. This upgrade allows the user to fill a
dpa_classifier table with MIXED (both) "return-to-keygen" AND
simple action entries.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>