dpa_offload/dpa_ipsec: Remove IPSec special manip chain before the node
Switch the order of operations in the process of releasing the IPSec
special manip chain from an IPSec SA. Originally the dpa_ipsec driver
would attempt to release the low level header manip node and afterwards
remove the chain from the intermediate (DPAA offloading) driver. This
is incorrect as it would bypass the protection that dpa_classifier
is offering by knowing if the header manip chain is still in use by other
entities. This update reverses the order of operations and removes first
the header manip chain from the dpa_classifier and only afterwards,
in case of success, it removes the header manip node from the low level
driver (FMD).

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>