armv8: aarch32: Allow RAM to be mapped for LayerScape SoC
This patch is based on Kernel v4.1. As in v4.1, memremap() is not
introduced and the WARN() check is not relaxed to allow MT_MEMORY_RW
mappings of pfn_valid() pages, this patch is needed as a workaround
for the current problem.

For the latest kernel in upstream, this patch is not needed anymore.

Signed-off-by: Alison Wang <alison.wang@nxp.com>