staging: fsl-dpaa2/eth: Flow affinity on IP forwarding
The driver egress function chooses a FQ to send the frames based on
the current processor ID. The network stack itself sets a mapping
field in the skb based on many things - the default one being a hash
on packet fiels, which the current driver ignores.

This patch saves the ingress frame flow affinity information in the
skb. In case of forwarded frames, this info will then be used for Tx
and Tx confirmation hardware queue selection, ensuring all processing
of the given frame is done on a single core.

This brings a ~5% performance improvement for IP forwarding between
two DPNI interfaces.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@nxp.com>
Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>